update Otto's Expedition data nightly (#32)
this PR updates the nightly Otto's Expedition GitHub Action to upload
new synthetic data to the GCS bucket. the logic is:

- every night or on PRs with the `run_datagen_on_pr` label...
- alternatively, we could run it on PRs only when given the label (and
remove it as part of the action)
- generate the last 7 days of data
- `rsync` it to the bucket (generally overwriting the last 6 days of
data)

steps taken:

- create service account
- give write permissions on GCS bucket to service account
- add service account secret JSON file as a GitHub action repo secret
- add GCP project ID as a GitHub action repo variable 

this may be fine as-is, or we might want more sophisitication. or just
generate 1 day of data back, but we risk missing data if GitHub Actions
don't run for whatever reason. also feature not a bug -- easier backfill
demos? "the underlying data for the last 6 days changed, let's backfill
and see the differences"